+++ src/CognitiveServices/SpeechRecognition.ts
@@ -2,6 +2,16 @@ import { Speech, Func, Action } from '../SpeechModule'
 import * as konsole from '../Konsole';
 import * as CognitiveSpeech from 'microsoft-speech-browser-sdk/Speech.Browser.Sdk'
 
+export interface ISpeechContextDgiGroup {
+    Type: string;
+    Hints?: { ReferenceGrammar: string };
+    Items?: { Text: string }[];
+}
+
+export interface ISpeechContext {
+    dgi: { Groups: ISpeechContextDgiGroup[] };
+}
+
 export interface ICognitiveServicesSpeechRecognizerProperties {
     locale?: string,
     subscriptionKey?: string,
@@ -20,6 +30,7 @@ export class SpeechRecognizer implements Speech.ISpeechRecognizer {
     public referenceGrammarId: string;
 
     private actualRecognizer: any = null;
+    private grammars: string[] = null;
     private properties: ICognitiveServicesSpeechRecognizerProperties;
 
     constructor(properties: ICognitiveServicesSpeechRecognizerProperties = {}) {
@@ -67,6 +78,10 @@ export class SpeechRecognizer implements Speech.ISpeechRecognizer {
     }
 
     public warmup() {
+    }
+
+    public setGrammars(grammars: string[]) {
+        this.grammars = grammars;
     }
 
     public async startRecognizing() {
@@ -118,21 +133,23 @@ export class SpeechRecognizer implements Speech.ISpeechRecognizer {
             }
         }
 
-        let speechContext = null;
+        const speechContext: ISpeechContext = { dgi: { Groups: [] } };
+
         if (this.referenceGrammarId) {
-            speechContext = JSON.stringify({
-                dgi: {
-                    Groups: [
-                        {
-                            Type: "Generic",
-                            Hints: { ReferenceGrammar: this.referenceGrammarId }
-                        }
-                    ]
-                }
+            speechContext.dgi.Groups.push({
+                Type: 'Generic',
+                Hints: { ReferenceGrammar: this.referenceGrammarId }
             });
         }
 
-        return this.actualRecognizer.Recognize(eventhandler, speechContext);
+        if (this.grammars) {
+            speechContext.dgi.Groups.push({
+                Type: 'Generic',
+                Items: this.grammars.map(grammar => ({ Text: grammar }))
+            });
+        }
+
+        return this.actualRecognizer.Recognize(eventhandler, JSON.stringify(speechContext));
     }
 
     public speechIsAvailable(){